fix: move service into components/toast
