+++ angular/bootstrap/src/components/toast/index.ts
@@ -1,3 +1,4 @@
 export * from './toast.component';
 export * from './toast.gen';
 export * from './toaster.component';
+export * from './toaster.service';
