+++ angular/bootstrap/src/components/toast/toaster.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import type {ToastProps} from '../components/toast/toast.gen';
+import type {ToastProps} from './toast.gen';
 import {ToasterService as headlessToaster} from '@agnos-ui/angular-headless';
 
 @Injectable({
