+++ angular/headless/src/components/toast/index.ts
@@ -1,0 +1,2 @@
+export * from '@agnos-ui/core/components/toast';
+export * from './toaster.service';
