+++ angular/headless/src/components/toast/toaster.service.ts
@@ -1,9 +1,12 @@
 import {type ToastProps} from '@agnos-ui/core/components/toast';
-import type {ToasterToast, ToasterProps} from '@agnos-ui/core/services/toaster';
-import {Toaster as CoreToaster} from '@agnos-ui/core/services/toaster';
-import {toAngularSignal, toAngularWritableSignal} from '../generated';
+import type {ToasterToast, ToasterProps} from '@agnos-ui/core/components/toast';
+import {Toaster as CoreToaster} from '@agnos-ui/core/components/toast';
+import {toAngularSignal, toAngularWritableSignal} from '../../generated';
 import type {Signal, WritableSignal} from '@angular/core';
 import {inject, Injectable, InjectionToken} from '@angular/core';
+
+import type {ToastPositions} from '@agnos-ui/core/components/toast';
+export type {ToastPositions};
 
 export const ToastPropsToken = new InjectionToken<ToasterProps>('ToasterProps');
 
@@ -16,12 +19,12 @@ export const ToastPropsToken = new InjectionToken<ToasterProps>('ToasterProps');
 	providedIn: 'root',
 })
 export class ToasterService<Props extends Partial<ToastProps>> {
-	readonly optionsCore = inject(ToastPropsToken, {optional: true});;
+	readonly optionsCore = inject(ToastPropsToken, {optional: true});
 	readonly #toaster: CoreToaster<Props> = new CoreToaster<Props>(this.optionsCore ?? undefined);
-	readonly toasts: Signal<ToasterToast<Props>[]> = toAngularSignal(this.#toaster.toasts);;
-	readonly options: WritableSignal<ToasterProps> = toAngularWritableSignal(this.#toaster.options);;
+	readonly toasts: Signal<ToasterToast<Props>[]> = toAngularSignal(this.#toaster.toasts);
+	readonly options: WritableSignal<ToasterProps> = toAngularWritableSignal(this.#toaster.options);
 	readonly addToast = this.#toaster.addToast;
 	readonly removeToast = this.#toaster.removeToast;
 	readonly eventsDirective = this.#toaster.eventsDirective;
-	readonly closeAll = this.#toaster.closeAll
+	readonly closeAll = this.#toaster.closeAll;
 }
