+++ angular/headless/src/index.ts
@@ -2,4 +2,4 @@
  * Public API Surface of @agnos-ui/angular-headless
  */
 export * from './generated';
-export * from './services';
+export * from './components/toast';
