+++ core-bootstrap/src/scss/toaster.scss
@@ -1,8 +1,7 @@
-
 .au-toaster {
-    display: flex;
+	display: flex;
 }
 
 .au-toaster-container {
-    padding: 1rem;
+	padding: 1rem;
 }
