+++ core/src/components/toast/index.ts
@@ -1,1 +1,2 @@
 export * from './toast';
+export * from './toaster';
