+++ react/bootstrap/src/components/toast/index.ts
@@ -1,2 +1,3 @@
 export * from './toast';
 export * from './toast.gen';
+export * from './toasterProvider';
