+++ svelte/bootstrap/src/components/toast/toasterService.ts
@@ -1,0 +1,3 @@
+import type {ToastProps} from './toast.gen';
+import {Toaster as headlessToaster} from '@agnos-ui/svelte-headless/components/toast';
+export class ToasterService extends headlessToaster<Partial<ToastProps>> {}
