+++ svelte/bootstrap/src/index.ts
@@ -11,6 +11,4 @@ export * from './components/slider';
 export * from './components/toast';
 export * from './components/tree';
 
-export * from './services/toaster';
-
 export * from './generated';
