+++ svelte/bootstrap/src/types.ts
@@ -1,1 +1,2 @@
+export * from '@agnos-ui/svelte-headless/types';
 export * from '@agnos-ui/core-bootstrap/types';
