+++ svelte/headless/src/components/toast/index.ts
@@ -1,0 +1,3 @@
+export * from '@agnos-ui/core/components/toast';
+export {Toaster} from './toasterService';
+export * from './toasterService';
