+++ svelte/headless/src/components/toast/toasterService.ts
@@ -1,10 +1,9 @@
-import {type ToastProps} from '@agnos-ui/core/components/toast';
-import type {ToasterToast, ToasterProps, ToastPositions} from '@agnos-ui/core/services/toaster';
-import {Toaster as CoreToaster} from '@agnos-ui/core/services/toaster';
+import type {Directive} from '../../generated';
+import type {ToastProps, ToasterToast, ToasterProps, ToastPositions} from '@agnos-ui/core/components/toast';
+import {Toaster as CoreToaster} from '@agnos-ui/core/components/toast';
 import {fromStore} from 'svelte/store';
-import type {Directive} from '../generated';
+import {defaultToasterProps} from '@agnos-ui/core/components/toast';
 
-import {defaultToasterProps} from '@agnos-ui/core/services/toaster';
 export {defaultToasterProps, ToasterProps};
 export type {ToastPositions};
 
@@ -18,7 +17,7 @@ export class Toaster<Props extends Partial<ToastProps>> {
 	readonly #toaster: CoreToaster<Props>;
 	readonly toasts: {current: ToasterToast<Props>[]};
 	readonly options: {current: ToasterProps};
-	readonly addToast: (props: Props) => void;
+	readonly addToast: (props: Props) => number;
 	readonly removeToast: (id: number) => void;
 	readonly eventsDirective: Directive<number>;
 	readonly closeAll: () => void;
